test(action-sheet): migrate isOpen test
